Add WebServices

+++ src/export/exportdata.py
@@ -4,6 +4,8 @@ Auteur: Bruno DELATTRE
 Date : 09/12/2016
 """
 
+import os
+
 from dal import dal_dht22
 from lib import com_config, com_logger, com_sqlite
 
@@ -22,13 +24,32 @@ class ExportData:
     def export(self):
         self.logger.info('Get data from database')
         dal = dal_dht22.DAL_DHT22(self.connexion, self.cursor)
-        rows = dal.get_dht22()
-        
+        lastdatarows = dal.get_lastdata()
         self.logger.info('Write file')
-        # TODO read last line and send date to req
-        file = open(self.config['EXPORT']['file'], 'w')
-        file.write('Date,Capteur,Température,Humidité\n')
+        
+        # Read last line and send date to req
+        if os.path.isfile(self.config['EXPORT']['lastexport']):
+            file = open(self.config['EXPORT']['lastexport'], 'r')
+            lastdate = file.read()
+            file.close()
+            rows = dal.get_dht22(str(lastdate))
+        else:
+            lastdate = '2000-01-01 00:00:00'
+            rows = dal.get_dht22(lastdate)
+        
+        if not os.path.isfile(self.config['EXPORT']['file']):
+            # Creaate Data file
+            file = open(self.config['EXPORT']['file'], 'w')
+            file.write('Date,Capteur,Temperature,Humidite\n')
+            file.close()
+        
+        file = open(self.config['EXPORT']['file'], 'a+')
         for row in rows:
             file.write(row[0] + ',' + row[1] + ',' + str(row[2]) + ',' + str(row[3]) + '\n')
         file.close()
         self.logger.info('Export done: ' + self.config['EXPORT']['file'])
+        
+        # Create ast entry
+        file = open(self.config['EXPORT']['lastexport'], 'w')
+        file.write(lastdatarows[0][0])
+        file.close()
